github_etl_docker/requirements.txt
requests
psycopg2-binary
python-dotenv
pandas
sqlalchemy